improve header, add scss
