+++ package.json
@@ -48,6 +48,7 @@
     "react-query": "^3.34.16",
     "react-share": "^4.4.0",
     "react-toastify": "^8.2.0",
+    "sass": "^1.49.9",
     "tailwindcss": "^3.0.23",
     "toastify": "^2.0.1",
     "truncate": "^3.0.0"
